requirements.txt
absl-py==1.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
black==23.11.0
blinker==1.4
cachetools==5.2.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.3.1
cloudpickle==3.0.0
colorama==0.4.3
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
datasets==2.15.0
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
dill==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
flatbuffers==23.1.4
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
httplib2==0.14.0
huggingface-hub==0.19.4
hydra-core==1.3.2
hyperlink==19.0.0
idna==2.8
imageio==2.33.1
imbalanced-learn==0.11.0
imgviz==1.7.4
importlib-metadata==6.0.0
importlib-resources==6.1.1
incremental==16.10.1
iopath==0.1.9
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
Jinja2==2.10.1
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.6.0
jupyter-core==5.5.0
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
libclang==15.0.6.1
lightgbm==4.1.0
linecache2==1.0.0
llvmlite==0.41.1
macaroonbakery==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.4
matplotlib-inline==0.1.6
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.10.4
networkx==3.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
omegaconf==2.3.0
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
panopticapi==0.1
parso==0.8.3
pathspec==0.11.2
patsy==0.5.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.0.0
plotly==5.18.0
plotly-express==0.4.1
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==3.19.6
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycocotools==2.0.7
pycups==1.9.73
pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.1.2
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.0
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==2.3.1
service-identity==18.1.0
shap==0.43.0
shapely==2.0.2
simplejson==3.16.0
six==1.16.0
slicer==0.0.7
sos==4.5.6
ssh-import-id==5.10
stack-data==0.6.3
statsmodels==0.13.5
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
testresources==2.0.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tornado==6.4
tqdm==4.66.1
traceback2==1.4.0
traitlets==5.14.0
triton==2.1.0
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
ultralytics==8.0.221
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.12
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.3
zipp==3.17.0
zope.interface==4.7.1
